# swagger.yaml
openapi: 3.0.3
info:
  title: EPOS Backoffice RESTful APIs
  description: This is the Backoffice RESTful APIs Swagger page.
  contact:
    email: apis@lists.epos-ip.org
  license:
    name: MIT License
    url: https://epos-ci.brgm.fr/epos/WebApi/raw/master/LICENSE
  version: 2.0.0
servers:
  - url: http://localhost:8080/api/backoffice/v1
    description: Inferred Url
paths:
...
  /dataproduct:
    post:
      tags:
        - V3 - DataProduct
      summary: "Create DataProduct instance"
      description: ""
      operationId: "dataproductpost"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/dataproduct'
      responses:
        '201':
          description: Created.
        '204':
          description: No content.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    put:
      tags:
        - V3 - DataProduct
      summary: "Update DataProduct"
      description: ""
      operationId: "dataproductputstate"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/dataproduct'
      responses:
        '200':
          description: ok.
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /dataproduct/{instance_id}:
    get:
      tags:
        - V3 - DataProduct
      summary: "Retrive DataProduct"
      description: ""
      operationId: "dataproductget"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Instance Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dataproduct'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    delete:
      tags:
        - V3 - DataProduct
      summary: "Delete DataProduct"
      description: ""
      operationId: "dataproductdelete"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dataproduct'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
...
  /contactpoint:
    post:
      tags:
        - V3 - ContactPoint
      summary: "Create ContactPoint instance"
      description: ""
      operationId: "contactpointpost"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/contactpoint'
      responses:
        '201':
          description: Created.
        '204':
          description: No content.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    put:
      tags:
        - V3 - ContactPoint
      summary: "Update ContactPoint"
      description: ""
      operationId: "contactpointputstate"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/contactpoint'
      responses:
        '200':
          description: ok.
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /contactpoint/{instance_id}:
    get:
      tags:
        - V3 - ContactPoint
      summary: "Retrive ContactPoint"
      description: ""
      operationId: "contactpointget"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Instance Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contactpoint'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    delete:
      tags:
        - V3 - ContactPoint
      summary: "Delete ContactPoint"
      description: ""
      operationId: "contactpointdelete"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contactpoint'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
...
components:
  schemas:
...
    contactpoint:
      required:
        - UID
      properties:
        UID:
          type: string
          example: uid
          description: 'This property contains the main identifier for the Contact Point'
        email:
          type: array
          items:
            type: string
            example: 'email'
          description: 'This property refers to the Contact Point Email.'
        language:
          type: array
          items:
            type: string
            example: 'language'
          description: 'This property contains information about the language used by the ContactPoint. Please use one of the language codes from the IETF BCP 47 standard.'
        role:
          type: string
          example: 'role'
          description: 'This property refers to the Contact Point role.'
        person:
          type: string
          example: 'person UID'
          description: 'This property refers to the person who is the Contact Point.'
        telephone:
          type: array
          items:
            type: string
            example: 'telephone'
          description: 'This property refers to the Contact Point Telephone.'
...
    dataproduct:
      required:
        - UID
        - description
        - title
      type: object
      properties:
        UID:
          type: string
          example: uid
          description: 'This property contains the main identifier for the Data Product, e.g. the URI or other unique identifier in the context of the Catalogue.'
        accessRight:
          type: array
          items:
            type: string
            example: 'access right'
          description: 'This property refers to information that indicates whether the Data Product is open data, has access restrictions or is not public.'
        accrualPeriodicity:
          type: string
          example: 'accrual periodicity'
          description: 'This property refers to the frequency at which the Data Product is updated.'
        category:
          type: array
          items:
            type: string
            example: 'category'
          description: 'This property refers to a category of the Data Product. A Data Product may be associated with multiple categories.'
        contactPoint:
          type: array
          items:
            type: string
            example: 'contactPoint UID'
          description: 'This property contains contact information (i.e. Role) that can be used for sending comments about the Data Product.'
        created:
          type: string
          format: date-time
          description: 'This property contains the date on which the Data Product was created.'
        description:
          type: array
          items:
            type: string
            example: 'description'
          description: 'This property contains a free-text account of the Data Product. This property can be repeated for parallel language versions of the description.'
        distribution:
          type: array
          items:
            type: string
            example: 'distribution UID'
          description: 'This property links the Data Product to an available Distribution.'
        hasPart:
          type: array
          items:
            type: string
            example: 'dataproduct UID'
          description: 'This property refers to a related Data Product that is part of the described Data Product.'
        identifier:
          type: array
          items:
            $ref: '#/components/schemas/identifier'
          description: 'This property refers to a secondary identifier of the Data Product, such as MAST/ADS, DataCite, DOI, EZID or W3ID.'
        isPartOf:
          type: array
          items:
            type: string
            example: 'dataproduct'
          description: 'This property refers to a related Data Product in which the described Data Product is physically or logically included.'
        issued:
          type: string
          format: date-time
          description: 'This property contains the date of formal issuance (e.g., publication) of the Data Product.'
        keywords:
          type: string
          example: 'keyword1, keyword2'
          description: 'This property contains a keyword or tag describing the Data Product. Multiple entries in a keywords list are typically delimited by commas.'
        modified:
          type: string
          format: date-time
          description: 'This property contains the most recent date on which the Data Product was changed or modified.'
        provenance:
          type: array
          items:
            type: string
            example: 'provenance'
          description: 'This property contains a statement about the lineage of a Data Product'
        publisher:
          type: array
          items:
            type: string
            example: 'organization'
          description: 'This property refers to an entity (organisation) responsible for making the Data Product available.'
        relation:
          type: array
          items:
            type: string
            example: 'epos resource'
          description: 'It represents the link to another Epos resource.'
        spatialExtent:
          type: array
          items:
            $ref: '#/components/schemas/location'
          description: 'This property refers to a geographic region that is covered by the Data Product.'
        temporalExtent:
          type: array
          items:
            $ref: '#/components/schemas/periodOfTime'
          description: 'This property refers to a temporal period (i.e. startDate, endDate) that the Data Product covers.'
        title:
          type: array
          items:
            type: string
            example: title
          description: 'This property contains a name given to the Data Product. This property can be repeated for parallel language versions of the name.'
        type:
          type: string
          example: type
          description: 'This property refers to the type of the Data Product. A controlled vocabulary for the values has not been established.'
        versionInfo:
          type: string
          example: '1.0'
          description: 'This property contains a version number or other version designation of the Data Product.'
...
    location:
      type: object
      properties:
        latitude:
          type: string
          example: 'coordinate latitude'
        longitude:
          type: string
          example: 'coordinate longitude'
        geobbox:
          type: array
          items:
            type: string
            example: 'geobbox'
    periodOfTime:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
...
    identifier:
      type: object
      required:
        - identifier
        - type
      properties:
        type:
          type: string
          example: 'DDSS-ID'
          description: 'This property refers the identifier scheme referenced by its datatype (e.g., DDSS-ID, DOI, etc.).'
        identifier:
          type: string
          example: 'id'
          description: 'This property contains a string that is an identifier in the context of the identifier scheme referenced by its datatype'
...
  securitySchemes:
    aai:
      type: apiKey
      in: header
      name: Authorization